+++ dockerfiles/resources/sql/schema.sql
@@ -62,3 +62,22 @@ create table proposals
     scores_updated bigint,
     votes          bigint
 );
+
+create table registered_events
+(
+    id         bigserial
+        primary key,
+    created_at timestamp with time zone,
+    updated_at timestamp with time zone,
+    deleted_at timestamp with time zone,
+    type       text,
+    event      text,
+    type_id    text
+);
+
+create index idx_registered_events_deleted_at
+    on registered_events (deleted_at);
+
+alter table registered_events
+    add constraint idx_unique_registered_events
+        unique (type, type_id, event);