+++ internal/app.go
@@ -14,6 +14,7 @@ import (
 	"github.com/goverland-labs/core-storage/internal/communicate"
 	"github.com/goverland-labs/core-storage/internal/config"
 	"github.com/goverland-labs/core-storage/internal/dao"
+	"github.com/goverland-labs/core-storage/internal/events"
 	"github.com/goverland-labs/core-storage/internal/proposal"
 	"github.com/goverland-labs/core-storage/pkg/health"
 	"github.com/goverland-labs/core-storage/pkg/prometheus"
@@ -82,11 +83,8 @@ func (a *Application) initDB() error {
 	}
 	ps.SetMaxOpenConns(a.cfg.DB.MaxOpenConnections)
 
-	a.db = db
-
-	// todo: remove automigrations
-	dao.AutoMigrate(db)
-	proposal.AutoMigrate(db)
+	// todo: remove debug after testing
+	a.db = db.Debug()
 
 	return err
 }
@@ -138,8 +136,14 @@ func (a *Application) initDao(nc *nats.Conn, pb *communicate.Publisher) error {
 }
 
 func (a *Application) initProposal(nc *nats.Conn, pb *communicate.Publisher) error {
+	erRepo := events.NewRepo(a.db)
+	erService, err := events.NewService(erRepo)
+	if err != nil {
+		return fmt.Errorf("new events service: %w", err)
+	}
+
 	repo := proposal.NewRepo(a.db)
-	service, err := proposal.NewService(repo, pb)
+	service, err := proposal.NewService(repo, pb, erService)
 	if err != nil {
 		return fmt.Errorf("proposal service: %w", err)
 	}
@@ -150,6 +154,9 @@ func (a *Application) initProposal(nc *nats.Conn, pb *communicate.Publisher) err
 	}
 
 	a.manager.AddWorker(process.NewCallbackWorker("proposal-consumer", cs.Start))
+
+	vw := proposal.NewVotingWorker(service)
+	a.manager.AddWorker(process.NewCallbackWorker("voting-worker", vw.Start))
 
 	return nil
 }