+++ internal/proposal/repo.go
@@ -2,6 +2,7 @@ package proposal
 
 import (
 	"fmt"
+	"time"
 
 	"gorm.io/gorm"
 )
@@ -34,4 +35,15 @@ func (r *Repo) GetByID(id string) (*Proposal, error) {
 	}
 
 	return &p, nil
+}
+
+// todo: think about limits, add pagination or cursor
+func (r *Repo) GetAvailableForVoting(window time.Duration) ([]*Proposal, error) {
+	var items []*Proposal
+	err := r.db.Raw("select * from proposals p where p.end > ?", time.Now().Add(-window).Unix()).Find(&items).Error
+	if err != nil {
+		return nil, fmt.Errorf("find active: %w", err)
+	}
+
+	return items, nil
 }